:recycle: refactor: 패키지 구조 변경
- controller, service, repository 패키지 하위의 존재하는 세부 패키지 제거
