+++ airdnb/src/test/java/com/example/airdnb/repository/user/UserRepositoryTest.java
@@ -4,6 +4,7 @@ import static org.assertj.core.api.Assertions.assertThat;
 
 import com.example.airdnb.domain.user.User;
 import com.example.airdnb.domain.user.User.Role;
+import com.example.airdnb.repository.UserRepository;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.orm.jpa.DataJpaTest;
